+++ eduVPN.Client/App.xaml.cs
@@ -47,6 +47,7 @@ namespace eduVPN.Client
             base.OnStartup(e);
 
             // Set client-specific settings.
+            eduVPN.Properties.SettingsEx.Default.RegistryKeyPath = @"SOFTWARE\SURF\eduVPN";
             eduVPN.Properties.Settings.Default.SelfUpdateBundleId = "{EF5D5806-B90B-4AA3-800A-2D7EA1592BA0}";
             eduVPN.Properties.Settings.Default.ClientId = "org.eduvpn.app";
             eduVPN.Properties.Settings.Default.ClientTitle = Client.Resources.Strings.ConnectWizardTitle;