+++ eduVPN.Views/App.cs
@@ -44,29 +44,6 @@ namespace eduVPN.Views
             //System.Threading.Thread.CurrentThread.CurrentCulture = culture;
             //System.Threading.Thread.CurrentThread.CurrentUICulture = culture;
 
-            // If app.local.config file is present, read configuration from it.
-            var appConfigFile = AppDomain.CurrentDomain.SetupInformation.ConfigurationFile;
-            var localConfigFile = Path.Combine(Path.GetDirectoryName(appConfigFile), Path.GetFileNameWithoutExtension(appConfigFile) + ".local.config");
-            if (File.Exists(localConfigFile))
-            {
-                AppDomain.CurrentDomain.SetData("APP_CONFIG_FILE", localConfigFile);
-
-                typeof(ConfigurationManager)
-                    .GetField("s_initState", BindingFlags.NonPublic | BindingFlags.Static)
-                    .SetValue(null, 0);
-
-                typeof(ConfigurationManager)
-                    .GetField("s_configSystem", BindingFlags.NonPublic | BindingFlags.Static)
-                    .SetValue(null, null);
-
-                typeof(ConfigurationManager)
-                    .Assembly.GetTypes()
-                    .Where(x => x.FullName == "System.Configuration.ClientConfigPaths")
-                    .First()
-                    .GetField("s_current", BindingFlags.NonPublic | BindingFlags.Static)
-                    .SetValue(null, null);
-            }
-
             try
             {
                 // Test load the user settings.