+++ CHANGES.md
@@ -4,6 +4,7 @@
 
 - GUI refinements
 - OpenVPN updated to 2.5.1-20210225
+- Customizing with .config.local files migrated to registry
 
 
 ## [1.255.3](https://github.com/Amebis/eduVPN/compare/1.255.2...1.255.3) (2021-02-18)